chore(e2e): optimize and clean tests
