+++ packages/web-components/tests/e2e-storybook/cypress/integration/card/card.e2e.js
@@ -77,31 +77,6 @@ const _tests = {
       capture: 'viewport',
     });
   },
-  checkTextRenders: () => {
-    it('should render eyebrow content', () => {
-      cy.get(_selectors.eyebrow).then($eyebrow => {
-        expect($eyebrow).not.to.be.empty;
-      });
-    });
-
-    it('should render heading content', () => {
-      cy.get(_selectors.heading).then($heading => {
-        expect($heading).not.to.be.empty;
-      });
-    });
-
-    it('should render body copy content', () => {
-      cy.get(_selectors.copy).then($copy => {
-        expect($copy).not.to.be.empty;
-      });
-    });
-
-    it('should render footer content', () => {
-      cy.get(_selectors.footer).then($footer => {
-        expect($footer).not.to.be.empty;
-      });
-    });
-  },
   checkClickableCard: pictogram => {
     if (pictogram) {
       it('should check for link', () => {
@@ -144,55 +119,10 @@ const _tests = {
         .focus();
     });
   },
-  checkNonClickableCard: () => {
-    it('should not respond to a click ', () => {
-      let initialLocation;
-      cy.location('href')
-        .then(location => {
-          initialLocation = location;
-        })
-        .get(_selectorBase)
-        .click()
-        .location('href')
-        .then(location => {
-          expect(location).to.equal(initialLocation);
-        });
-    });
-  },
   checkImageRenders: (path, groupId) => {
     it('should render with image', () => {
       cy.visit(`${path}&knob-Add%20image:${groupId}=true`);
       cy.get(_selectors.image).should('have.length', 1);
-      cy.takeSnapshots();
-    });
-  },
-  checkTagGroupRenders: (path, groupId) => {
-    it('should render with tag group', () => {
-      cy.visit(`${path}&knob-Add%20tags:${groupId}=true`);
-      cy.get(_selectors.tagGroup).should('have.length', 1);
-      cy.takeSnapshots();
-    });
-  },
-  checkOutlineRenders: path => {
-    it('should render with outline', () => {
-      cy.visit(path);
-      cy.get(_selectorBase).should('have.attr', 'border');
-      // converted HEX var(--cds-ui-03, #e0e0e0) to RGB
-      if (Cypress.browser.name !== 'firefox') {
-        cy.get(_selectorBase)
-          .shadow()
-          .find('.bx--card')
-          .should('have.css', 'border')
-          .and('equal', '1px solid rgb(224, 224, 224)');
-      }
-
-      cy.get(_selectorBase).should('have.attr', 'color-scheme', 'light');
-      // converted HEX var(--cds-ui-02, #ffffff) to RGB
-      cy.get(_selectorBase)
-        .shadow()
-        .find('.bx--card__wrapper')
-        .should('have.css', 'background-color')
-        .and('equal', 'rgb(255, 255, 255)');
       cy.takeSnapshots();
     });
   },
@@ -212,20 +142,12 @@ const _tests = {
     if (position === 'top') {
       it('should check for pictogram at the top', () => {
         cy.get('dds-card').should('have.attr', 'pictogram-placement', 'top');
-        cy.get('dds-card svg').then($content => {
-          expect($content[0].getBoundingClientRect().top).to.equal(32);
-          expect($content[0].getBoundingClientRect().bottom).to.equal(80);
-        });
       });
     } else {
       it('should check for pictogram at the bottom with text showing on hover', () => {
         cy.visit(`/${_pathPictogram}&knob-Pictogram%20position:${groupId}=bottom`);
 
         cy.get('dds-card').should('have.attr', 'pictogram-placement', 'bottom');
-        cy.get('dds-card svg').then($content => {
-          expect($content[0].getBoundingClientRect().top).to.equal(186);
-          expect($content[0].getBoundingClientRect().bottom).to.equal(234);
-        });
 
         cy.get('dds-card').then($el => {
           const sheets = $el[0].shadowRoot.adoptedStyleSheets;
@@ -254,13 +176,10 @@ describe('dds-card | default (desktop)', () => {
     cy.viewport(1280, 780);
   });
 
-  _tests.checkTextRenders();
   _tests.checkClickableCard();
   _tests.checkTabbableCard();
   _tests.checkImageRenders(_path, groupId);
-  _tests.checkTagGroupRenders(_path, groupId);
   _tests.checkInverseRenders(`${_path}&knob-Card%20style:${groupId}=Inverse%20card`);
-  _tests.checkOutlineRenders(`${_path}&knob-Card%20style:${groupId}=Outlined%20card`);
   it('should render correctly in all themes', _tests.screenshotThemes);
   it('should check a11y', _tests.checkA11y);
 });
@@ -273,11 +192,8 @@ describe('dds-card | pictogram (desktop)', () => {
     cy.viewport(1280, 780);
   });
 
-  _tests.checkClickableCard(true);
-  _tests.pictogramPosition('top', groupId);
   _tests.pictogramPosition('bottom', groupId);
   _tests.checkInverseRenders(`${_pathPictogram}&knob-Card%20style:${groupId}=Inverse%20card`);
-  _tests.checkOutlineRenders(`${_pathPictogram}&knob-Card%20style:${groupId}=Outlined%20card`);
   it('should render correctly in all themes', _tests.screenshotThemes);
   it('should check a11y', _tests.checkA11y);
 });
@@ -290,11 +206,6 @@ describe('dds-card | static (desktop)', () => {
     cy.viewport(1280, 780);
   });
 
-  _tests.checkTextRenders();
-  _tests.checkClickableCard();
   _tests.checkImageRenders(_pathStatic, groupId);
-  _tests.checkTagGroupRenders(_pathStatic, groupId);
-  _tests.checkOutlineRenders(`${_pathStatic}&knob-Outlined%20card:${groupId}=true`);
-  it('should render correctly in all themes', _tests.screenshotThemes);
   it('should check a11y', _tests.checkA11y);
 });
