+++ packages/web-components/tests/e2e-storybook/cypress/integration/footer/footer.e2e.js
@@ -33,73 +33,6 @@ describe('dds-footer | default (desktop)', () => {
   it('should check a11y', () => {
     cy.checkAxeA11y();
   });
-
-  it('should have interactable url for IBM logo', () => {
-    cy.get('[data-autoid="dds--footer-logo"]').then($link => {
-      const url = $link.prop('href');
-      expect(url).not.to.be.empty;
-    });
-  });
-
-  it('should load locale modal', () => {
-    const localeButton = cy.get('[data-autoid="dds--locale-btn"]');
-    localeButton.click();
-
-    cy.screenshot();
-  });
-
-  it('should load the Americas region with its languages and locations', () => {
-    const localeButton = cy.get('[data-autoid="dds--locale-btn"]');
-    localeButton.click();
-
-    cy.get('dds-region-item[name="Americas"]').click();
-
-    cy.get('dds-locale-item[region="Americas"]').should('have.length', 35);
-
-    cy.takeSnapshots();
-  });
-
-  it('should be able to search with keywords for locations and languages', () => {
-    const localeButton = cy.get('[data-autoid="dds--locale-btn"]');
-    localeButton.click();
-
-    cy.get('[name="Americas"]').click();
-
-    cy.get('dds-locale-search')
-      .shadow()
-      .find('.bx--search-input')
-      .type('ca', {
-        force: true,
-      });
-
-    cy.get('dds-locale-item:not([hidden])')
-      .invoke('attr', 'country')
-      .should('eq', 'Canada');
-
-    cy.takeSnapshots();
-  });
-
-  it('should load clickable footer links', () => {
-    cy.get('dds-footer-nav-group')
-      .find('dds-footer-nav-item')
-      .each($link => {
-        const url = $link.prop('href');
-        expect(url).not.to.be.empty;
-      });
-
-    cy.takeSnapshots();
-  });
-
-  it('should load clickable legal links', () => {
-    cy.get('dds-legal-nav')
-      .find('dds-legal-nav-item')
-      .each($link => {
-        const url = $link.prop('href');
-        expect(url).not.to.be.empty;
-      });
-
-    cy.takeSnapshots();
-  });
 });
 
 describe('dds-footer | Default language only (desktop)', () => {
@@ -131,104 +64,5 @@ describe('dds-footer | Default language only (desktop)', () => {
     }
 
     cy.takeSnapshots();
-  });
-});
-
-describe('dds-footer | default (mobile)', () => {
-  beforeEach(() => {
-    cy.visit(`/${_pathDefault}`);
-    cy.viewport(320, 780);
-
-    cy.waitUntil(() => cy.get('[data-autoid="dds--footer-legal-nav"]').should('not.be.empty'));
-  });
-
-  it('should have interactable url for IBM logo', () => {
-    cy.get('[data-autoid="dds--footer-logo"]').then($link => {
-      const url = $link.prop('href');
-      expect(url).not.to.be.empty;
-    });
-  });
-
-  it('should load locale modal', () => {
-    const localeButton = cy.get('[data-autoid="dds--locale-btn"]');
-    localeButton.click();
-
-    cy.screenshot();
-  });
-
-  it('should load the Americas region with its languages and locations', () => {
-    const localeButton = cy.get('[data-autoid="dds--locale-btn"]');
-    localeButton.click();
-
-    cy.get('dds-region-item[name="Americas"]').click();
-
-    cy.get('dds-locale-item[region="Americas"]').should('have.length', 35);
-
-    cy.takeSnapshots();
-  });
-
-  it('should be able to search with keywords for locations and languages', () => {
-    const localeButton = cy.get('[data-autoid="dds--locale-btn"]');
-    localeButton.click();
-
-    cy.get('[name="Americas"]').click();
-
-    cy.get('dds-locale-search')
-      .shadow()
-      .find('.bx--search-input')
-      .type('ca', {
-        force: true,
-      });
-
-    cy.get('dds-locale-item:not([hidden])')
-      .invoke('attr', 'country')
-      .should('eq', 'Canada');
-
-    cy.takeSnapshots();
-  });
-
-  it('should load clickable footer links', () => {
-    cy.get('dds-footer-nav-group')
-      .find('dds-footer-nav-item')
-      .each($link => {
-        const url = $link.prop('href');
-        expect(url).not.to.be.empty;
-      });
-
-    cy.takeSnapshots('mobile');
-  });
-
-  it('should load clickable legal links', () => {
-    cy.get('dds-legal-nav')
-      .find('dds-legal-nav-item')
-      .each($link => {
-        const url = $link.prop('href');
-        expect(url).not.to.be.empty;
-      });
-
-    cy.takeSnapshots('mobile');
-  });
-});
-
-describe('dds-footer | Default language only (mobile)', () => {
-  beforeEach(() => {
-    cy.visit(`/${_pathDefaultLanguageOnly}`);
-    cy.viewport(320, 780);
-
-    cy.waitUntil(() => cy.get('[data-autoid="dds--footer-legal-nav"]').should('not.be.empty'));
-  });
-
-  it('should load language selector dropdown and be interactive', () => {
-    cy.get('dds-language-selector-mobile').should('have.length', 1);
-    cy.get('dds-language-selector-mobile')
-      .shadow()
-      .find('select.bx--select-input')
-      .select('Arabic / عربية');
-    cy.get('dds-language-selector-mobile')
-      .shadow()
-      .find('select.bx--select-input')
-      .should('have.value', 'Arabic / عربية');
-
-    cy.takeSnapshots('mobile');
   });
 });
