+++ packages/web-components/tests/e2e-storybook/cypress/integration/pictogram-item/pictogram-item.e2e.js
@@ -20,30 +20,6 @@ describe('dds-pictogram-item | Pictogram item (desktop)', () => {
     cy.checkAxeA11y();
   });
 
-  it('should load pictogram item and content', () => {
-    cy.visit(`/${_pathDefault}`);
-    cy.viewport(1280, 780);
-    cy.get('[data-autoid="dds--pictogram-item"]').should('have.length', 1);
-    cy.get('[data-autoid="dds--pictogram-item"] [data-autoid="dds--pictogram-item__pictogram"]').should('have.length', 1);
-    cy.get('[data-autoid="dds--pictogram-item"] [data-autoid="dds--content-item__heading"]').should('have.length', 1);
-
-    cy.takeSnapshots();
-  });
-
-  it('should check that the Link with icon is loaded and clickable', () => {
-    cy.visit(`/${_pathDefault}`);
-    cy.viewport(1280, 780);
-    const linkWithIcon = cy.get('[data-autoid="dds--pictogram-item"] [data-autoid="dds--link-with-icon"]');
-    linkWithIcon.should('have.length', 1);
-    linkWithIcon
-      .shadow()
-      .find('a')
-      .each($link => {
-        const url = $link.prop('href');
-        expect(url).not.to.be.empty;
-      });
-  });
-
   it('should support customizable pictogram SVGs', () => {
     cy.visit(`/${_pathDefault}&knob-Pictogram%20(required)=Touch`);
     cy.viewport(1280, 780);
@@ -55,50 +31,5 @@ describe('dds-pictogram-item | Pictogram item (desktop)', () => {
     cy.visit(`/${_pathDefault}`);
     cy.viewport(1280, 780);
     cy.carbonThemesScreenshot();
-  });
-});
-
-describe('dds-pictogram-item | Pictogram item (mobile)', () => {
-  it('should load pictogram item and content', () => {
-    cy.visit(`/${_pathDefault}`);
-    cy.viewport(320, 780);
-    cy.get('[data-autoid="dds--pictogram-item"]').should('have.length', 1);
-    cy.get('[data-autoid="dds--pictogram-item"] [data-autoid="dds--pictogram-item__pictogram"]').should('have.length', 1);
-    cy.get('[data-autoid="dds--pictogram-item"] [data-autoid="dds--content-item__heading"]').should('have.length', 1);
-
-    cy.takeSnapshots('mobile');
-  });
-
-  it('should check that the Link with icon is loaded and clickable', () => {
-    cy.visit(`/${_pathDefault}`);
-    cy.viewport(320, 780);
-    const linkWithIcon = cy.get('[data-autoid="dds--pictogram-item"] [data-autoid="dds--link-with-icon"]');
-    linkWithIcon.should('have.length', 1);
-    linkWithIcon
-      .shadow()
-      .find('a')
-      .each($link => {
-        const url = $link.prop('href');
-        expect(url).not.to.be.empty;
-      });
-  });
-
-  it('should support customizable pictogram SVGs', () => {
-    cy.visit(`/${_pathDefault}&knob-Pictogram%20(required)=Touch`);
-    cy.viewport(320, 780);
-
-    cy.takeSnapshots('mobile');
-  });
-
-  it('should load correctly in all themes', () => {
-    cy.visit(`/${_pathDefault}`);
-    cy.viewport(320, 780);
-
-    cy.carbonThemesScreenshot(
-      {},
-      {
-        width: [320],
-      }
-    );
   });
 });
