+++ packages/web-components/tests/e2e-storybook/cypress/integration/pricing-table/with-subheaders.e2e.js
@@ -97,7 +97,7 @@ const _tests = (path = path) => [
   },
 ];
 
-describe('dds-pricing-table | with subheaders (desktop)', () => {
+xdescribe('dds-pricing-table | with subheaders (desktop)', () => {
   beforeEach(() => {
     cy.viewport(1280, 780);
   });
@@ -105,7 +105,7 @@ describe('dds-pricing-table | with subheaders (desktop)', () => {
   [...createTests(path), ..._tests(path)].forEach(test => test());
 });
 
-describe('dds-pricing-table | with subheaders (mobile)', () => {
+xdescribe('dds-pricing-table | with subheaders (mobile)', () => {
   beforeEach(() => {
     cy.viewport(320, 720);
   });
