+++ packages/web-components/tests/e2e-storybook/cypress/integration/structured-list/structured-list.e2e.js
@@ -33,34 +33,6 @@ describe('dds-structured-list | default (desktop)', () => {
   });
 });
 
-describe('dds-structured-list | default (mobile)', () => {
-  beforeEach(() => {
-    cy.visit(`${_pathRoot}--default`)
-      .injectAxe()
-      .viewport(320, 780);
-  });
-
-  it('should check a11y', () => {
-    cy.checkAxeA11y();
-  });
-
-  it('should have column headers', () => {
-    cy.get('dds-structured-list-header-row dds-structured-list-header-cell')
-      .should('exist')
-      .each(header_cell => {
-        cy.get(header_cell).should('not.be.empty');
-      });
-  });
-
-  it('should scroll to display all cells', () => {
-    cy.get('dds-structured-list-header-cell, dds-structured-list-cell').each(cell => {
-      cy.get(cell)
-        .scrollIntoView()
-        .should('be.visible');
-    });
-  });
-});
-
 describe('dds-structured-list | With Row Headers (desktop)', () => {
   beforeEach(() => {
     cy.visit(`${_pathRoot}--with-row-headers`)
@@ -78,34 +50,6 @@ describe('dds-structured-list | With Row Headers (desktop)', () => {
       .each(header_cell => {
         cy.get(header_cell).should('not.be.empty');
       });
-  });
-});
-
-describe('dds-structured-list | With Row Headers (mobile)', () => {
-  beforeEach(() => {
-    cy.visit(`${_pathRoot}--with-row-headers`)
-      .injectAxe()
-      .viewport(320, 780);
-  });
-
-  it('should check a11y', () => {
-    cy.checkAxeA11y();
-  });
-
-  it('should have row headers', () => {
-    cy.get('dds-structured-list-row dds-structured-list-header-cell')
-      .should('exist')
-      .each(header_cell => {
-        cy.get(header_cell).should('not.be.empty');
-      });
-  });
-
-  it('should scroll to display all cells', () => {
-    cy.get('dds-structured-list-header-cell, dds-structured-list-cell').each(cell => {
-      cy.get(cell)
-        .scrollIntoView()
-        .should('be.visible');
-    });
   });
 });
 
@@ -132,42 +76,15 @@ describe('dds-structured-list | With Subheaders (desktop)', () => {
   });
 });
 
-describe('dds-structured-list | With Subheaders (mobile)', () => {
+describe('dds-structured-list | With Complex Content', () => {
   beforeEach(() => {
-    cy.visit(`${_pathRoot}--with-subheaders`)
+    cy.visit(`${_pathRoot}--with-complex-content`)
       .injectAxe()
       .viewport(1280, 780);
   });
 
   it('should check a11y', () => {
     cy.checkAxeA11y();
-  });
-
-  it('should have Subheaders', () => {
-    cy.get('dds-structured-list-group')
-      .shadow()
-      .find('th')
-      .each(([header]) => {
-        const title = header.getRootNode().host.getAttribute('title');
-
-        cy.get(header).should('include.text', title);
-      });
-  });
-
-  it('should scroll to display all cells', () => {
-    cy.get('dds-structured-list-header-cell, dds-structured-list-cell').each(cell => {
-      cy.get(cell)
-        .scrollIntoView()
-        .should('be.visible');
-    });
-  });
-});
-
-describe('dds-structured-list | With Complex Content', () => {
-  beforeEach(() => {
-    cy.visit(`${_pathRoot}--with-complex-content`)
-      .injectAxe()
-      .viewport(1280, 780);
   });
 
   it('should inject tooltip with tooltip attribute', () => {
